Eliminating fileset and path binding
Just for Zhen to test.  Based on George's earlier patch.

From 4cbf6956fb753eac8c36597df5c873829d6839f5 Mon Sep 17 00:00:00 2001
From: George Amvrosiadis <gamvrosi@gmail.com>
Date: Sat, 15 Mar 2014 23:09:45 -0400
Subject: [PATCH 1/2] Disconnect 'path' from 'name' fileset attribute

Disconnect 'path' from 'name' attribute for files and filesets.
Previously, the 'path' denoted the dir where the fileset or file
resided, and 'name' would be appended to it as the file's name, or the
fileset's root directory. This patch amends 'path' attribute resolution
so that 'name' is not part of it, and is only used as an identifier.
This is consistent with the way 'name' is used is general.

Signed-off-by: George Amvrosiadis <gamvrosi@gmail.com>
---
 fileset.c        |   29 +----------------------------
 flowop_library.c |   15 ---------------
 2 files changed, 1 insertions(+), 43 deletions(-)